+++ h2integrate/core/finances.py
@@ -82,7 +82,6 @@ class ProFastComp(om.ExplicitComponent):
         land_cost = 0.0
 
         pf = ProFAST.ProFAST()
-
         if self.options["commodity_type"] == "hydrogen":
             pf.set_params(
                 "commodity",